Optimize webpack config file a bit
