+++ webpack.config.js
@@ -1,14 +1,10 @@
+require('babel/register');
 var webpack = require('webpack');
 var SimpleHtmlPrecompiler = require('simple-html-precompiler');
-var path = require('path');
-var React = require('react');
-require('babel/register');
-var Router = require('react-router');
-var routes = require('./components/routes.jsx');
+var Path = require('path');
 var paths = require('./scripts/paths.js');
-var englishStrings = require('./locales/en-US.json');
-var currencies = require('./data/currencies.js');
-var url = require('url');
+var routeFileContent = require('./scripts/route-file-content.js');
+
 require('habitat').load();
 
 module.exports = {
@@ -16,7 +12,7 @@ module.exports = {
   output: {
     filename: '[name].js',
     chunkFilename: '[id].chunk.js',
-    path: path.join('public')
+    path: Path.join('public')
   },
   resolve: {
     extensions: ['', '.js', '.jsx']
@@ -53,38 +49,8 @@ module.exports = {
     }),
     new webpack.ContextReplacementPlugin(/buffer/, require('buffer')),
     new SimpleHtmlPrecompiler(paths, function(outputPath, callback) {
-      Router.run(routes, outputPath, function(Handler) {
-        var values = {
-          currency: currencies.usd,
-          presets: currencies.usd.presets.single,
-          currencies: currencies,
-          amount: '',
-          frequency: 'single'
-        };
-        var index = React.createFactory(require('./pages/index.jsx'));
-        var page = React.createFactory(Handler);
-        var locale = url.parse(outputPath).pathname.split('/')[1];
-        var currentString, mergedStrings;
-
-        if (locale && require('./locales/index.js').indexOf(locale) !== -1) {
-          currentString = require('./locales/' + locale +'.json');
-          mergedStrings = Object.assign({}, englishStrings, currentString);
-          values = Object.assign({}, {locales : [locale], messages: mergedStrings}, values);
-        } else {
-          locale = 'en-US';
-          values = Object.assign({}, {locales : [locale], messages: englishStrings}, values);
-        }
-        callback(React.renderToStaticMarkup(index({
-          localeInfo: locale,
-          metaData: {
-            desc: values.messages.i_donated_to_mozilla,
-            title: values.messages.support_mozilla,
-            site_name: 'mozilla.org',
-            site_url: url.resolve(process.env.APPLICATION_URI, locale + '/'),
-            site_title: values.messages.give_to_mozilla
-          },
-          markup: React.renderToStaticMarkup(page(values))
-        })));
+      routeFileContent(outputPath, function(content) {
+        callback(content);
       });
     })
   ]
